+++ pheme/templatetags/charts/h_bar.py
@@ -19,7 +19,7 @@
 
 import itertools
 from typing import Dict
-from django.utils.safestring import mark_safe
+from django.utils.safestring import SafeText
 from pheme.templatetags.charts import (
     register,
     _severity_class_colors,
@@ -33,7 +33,7 @@ __ORIENTATION_LINE_TEMPLATE = """
 
 __ORIENTATION_LINE_TEXT_TEMPLATE = """
 <text class="orientation label" y="22" x="{x}" fill="#4C4C4D" dominant-baseline="central"
-style="text-anchor: middle;" width="{width}">{label}
+style="font-size:{font_size};font-family:{font_family};text-anchor: middle;" width="{width}">{label}
 </text>
 """
 __BAR_ELEMENT_TEMPLATE = """
@@ -43,22 +43,22 @@ __BAR_ELEMENT_TEMPLATE = """
 __BAR_TEMPLATE = """
 <g class="entry" transform="translate(0, {y})">
 <text class="label category" y="22" x="87.5" fill="#4C4C4D" dominant-baseline="central"
-style="text-anchor: middle;" width="175">{key}
+style="font-size:{font_size};font-family:{font_family};text-anchor: right;" width="{max_hostname_len}">{key}
 </text>
-<g transform="translate(175, 0)">
+<g transform="translate({max_hostname_len}, 0)">
 {orientation_lines}
 {bar_elements}
 </g>
 <g transform="translate(700, 0)">
 <text class="sum" y="22" x="10" fill="#4C4C4D" dominant-baseline="central"
-style="text-anchor: left;" width="100">{total}</text>
+style="font-size:{font_size};font-family:{font_family};text-anchor: left;" width="100">{total}</text>
 </g>
 </g>
 """
 __BAR_CHART_TEMPLATE = """
 <svg width="{width}" viewBox="0 0 {width} {height}" xmlns="http://www.w3.org/2000/svg">
 {bars}
-<g transform="translate(175, {bar_legend_y})">
+<g transform="translate({max_hostname_len}, {bar_legend_y})">
 {bar_legend}
 </g>
 {legend}
@@ -74,7 +74,9 @@ def h_bar_chart(
     bar_jump=44,
     orientation_basis=20,
     limit=10,
-) -> str:
+    font_family="Droid Sans",
+    font_size=10,
+) -> SafeText:
 
     """
     Returns a stacked horizontal bar chart in svg.
@@ -101,17 +103,20 @@ def h_bar_chart(
         orientation_basis - if higher than 0 than there will be a vertical line
             each orientation_basis.
         limit - limits the data by N first elements
+        font_family - the font family used within text elements
+        font_size - the font size used within text elements
     """
 
     data = dict(itertools.islice(chart_data.items(), limit))
     if not title_color:
         title_color = _severity_class_colors
-    max_width = svg_width - 175 - 100  # key and total placeholder
-    # highest sum of counts
     if not data.values():
-        return mark_safe("")
+        return SafeText("")
+    # multiply by 1.25 for kerning
+    max_hostname_len = max(len(k) for k in data.keys()) * font_size * 1.25
+    max_width = svg_width - max_hostname_len - 100  # key and total placeholder
+    # highest sum of counts
     max_sum = max([sum(list(counts.values())) for counts in data.values()])
-
     orientation_lines = ""
     orientation_labels = ""
 
@@ -119,10 +124,17 @@ def h_bar_chart(
         x_pos = i * orientation_basis / max_sum * max_width
         label = str(i * orientation_basis)
         orientation_lines += __ORIENTATION_LINE_TEMPLATE.format(
-            x=x_pos, height=bar_jump + 10
+            x=x_pos,
+            height=bar_jump + 10,
+            font_family=font_family,
+            font_size=font_size,
         )
         orientation_labels += __ORIENTATION_LINE_TEXT_TEMPLATE.format(
-            x=x_pos, width=len(label), label=label
+            x=x_pos,
+            width=len(label),
+            label=label,
+            font_family=font_family,
+            font_size=font_size,
         )
         return orientation_lines, orientation_labels
 
@@ -146,7 +158,12 @@ def h_bar_chart(
             color = title_color.get(category)
             width = count / max_sum * max_width
             elements += __BAR_ELEMENT_TEMPLATE.format(
-                x=element_x, width=width, color=color
+                x=element_x,
+                width=width,
+                color=color,
+                font_family=font_family,
+                font_size=font_size,
+                max_hostname_len=max_hostname_len,
             )
             element_x += width
 
@@ -156,6 +173,9 @@ def h_bar_chart(
             bar_elements=elements,
             orientation_lines=orientation_lines,
             total=sum(counts.values()),
+            font_family=font_family,
+            font_size=font_size,
+            max_hostname_len=max_hostname_len,
         )
     svg_element_lengths = len(data.keys()) * bar_jump + 50
     svg_chart = __BAR_CHART_TEMPLATE.format(
@@ -164,8 +184,9 @@ def h_bar_chart(
         bars=bars,
         bar_legend=orientation_labels,
         bar_legend_y=len(data.keys()) * bar_jump + 20,
-        legend=_build_legend(
-            svg_element_lengths + 10, svg_width, 16, title_color
-        ),
+        legend=_build_legend(svg_element_lengths + 10, svg_width, title_color),
+        font_family=font_family,
+        font_size=font_size,
+        max_hostname_len=max_hostname_len,
     )
-    return mark_safe(svg_chart)
+    return SafeText(svg_chart)